chore: rename application config file (#212)
* chore: rename application config file

* chore!: remove unused config messages
